feat(front): wip update table spacing and move helper
