+++ packages/front/svelte/components/SubventionsVersementsDashboard/helper.js
@@ -37,6 +37,7 @@ export const mapSubventionsAndVersements = ({ subventions, versements }) => {
         none: [],
         withKey: {}
     });
+
     const flatenElements = [
         ...Object.values(elementsGroupedByVersementKey.withKey),
         ...elementsGroupedByVersementKey.none
@@ -46,10 +47,17 @@ export const mapSubventionsAndVersements = ({ subventions, versements }) => {
     return uniformizedElements.sort(sortByDateAsc);
 };
 
-const getYearOfItem = item => {
-    if (item.isSub) return item.annee_demande;
-    if (item.periodeDebut) return new Date(item.periodeDebut).getFullYear();
-    return "";
+// Return year of subvention or versement as a string
+export const getYearOfElement = element => {
+    if (element.isSub) return getSubventionYear(element);
+    if (element.isVersement) return getVersementYear(element);
+};
+
+export const getSubventionYear = subvention => subvention.annee_demande;
+
+export const getVersementYear = versement => {
+    if (versement.periodeDebut) return new Date(versement.periodeDebut).getFullYear();
+    if (versement.dateOperation) return new Date(versement.dateOperation).getFullYear();
 };
 
 const groupByVersementKey = (acc, curr) => {
@@ -59,7 +67,9 @@ const groupByVersementKey = (acc, curr) => {
         return acc;
     }
 
-    const key = curr.versementKey + "-" + getYearOfItem(curr); // Discuter avec maxime pour faire cette manip cotée api
+    const year = getYearOfElement(curr);
+    const yearStr = year ? String(year) : "";
+    const key = `${curr.versementKey}-${yearStr}`; // Discuter avec maxime pour faire cette manip cotée api
 
     if (!acc.withKey[key]) acc.withKey[key] = [];
